refactor: 댓글 삭제시 id에 해당하는 댓글이 존재하지 않으면 예외를 터뜨리는 기능 추가

+++ core/src/main/java/org/cosmic/cafe/application/comment/CommentService.java
@@ -32,8 +32,9 @@ public class CommentService {
 
     public void deleteComment(String stringCommentId, UUID memberId) {
         UUID commentId = UUID.fromString(stringCommentId);
-        Optional<Comment> comment = commentRepository.findById(commentId);
-        comment.ifPresent(comment1 -> comment1.validateOwner(memberId));
+        Comment comment = commentRepository.findById(commentId)
+                        .orElseThrow(() -> new BadRequestException("해당 댓글이 존재하지 않습니다.", CommentErrorCode.NO_SUCH_COMMENT));
+        comment.validateOwner(memberId);
         commentRepository.deleteById(commentId);
     }
